requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
PyJWT[crypto]
cryptography
